Refactoring and moving legacy code that sends through Henvendelse to its own class
