+++ .github/workflows/preprod-pipeline.yml
@@ -5,7 +5,7 @@ on:
         branches:
             - 'preprod'
             - 'preprod-alt'
-            - 'sd1347-cleanup'
+            - 'legacy-arkivering-service'
         paths-ignore:
             - '**.md'
             - '**/**.md'
@@ -13,7 +13,7 @@ on:
         branches:
             - 'preprod'
             - 'preprod-alt'
-            - 'sd1347-cleanup'
+            - 'legacy-arkivering-service'
         paths-ignore:
             - '**.md'
             - '**/**.md'
