+++ sendsoknad-business/src/main/java/no/nav/sbl/dialogarena/soknadinnsending/business/service/soknadservice/InnsendtSoknadService.java
@@ -9,7 +9,6 @@ import no.nav.sbl.dialogarena.sendsoknad.domain.exception.SendSoknadException;
 import no.nav.sbl.dialogarena.sendsoknad.domain.kravdialoginformasjon.KravdialogInformasjon;
 import no.nav.sbl.dialogarena.sendsoknad.domain.kravdialoginformasjon.KravdialogInformasjonHolder;
 import no.nav.sbl.dialogarena.soknadinnsending.business.domain.InnsendtSoknad;
-import no.nav.sbl.dialogarena.soknadinnsending.business.service.Transformers;
 import no.nav.sbl.dialogarena.soknadinnsending.business.service.VedleggService;
 import no.nav.sbl.dialogarena.soknadinnsending.consumer.henvendelse.HenvendelseService;
 import org.apache.commons.lang3.LocaleUtils;
@@ -22,6 +21,7 @@ import java.util.Optional;
 import java.util.function.Predicate;
 
 import static java.util.stream.Collectors.toList;
+import static no.nav.sbl.dialogarena.sendsoknad.domain.Vedlegg.toInnsendingsvalg;
 import static no.nav.sbl.dialogarena.soknadinnsending.consumer.skjemaoppslag.SkjemaOppslagService.SKJEMANUMMER_KVITTERING;
 
 @Component
@@ -70,7 +70,7 @@ public class InnsendtSoknadService {
                 .map(xmlMetadata -> {
                     XMLVedlegg xmlVedlegg = (XMLVedlegg) xmlMetadata;
                     Vedlegg v = new Vedlegg()
-                            .medInnsendingsvalg(Transformers.toInnsendingsvalg(xmlVedlegg.getInnsendingsvalg()))
+                            .medInnsendingsvalg(toInnsendingsvalg(xmlVedlegg.getInnsendingsvalg()))
                             .medSkjemaNummer(xmlVedlegg.getSkjemanummer())
                             .medSkjemanummerTillegg(xmlVedlegg.getSkjemanummerTillegg())
                             .medNavn(TilleggsInfoService.lesTittelFraJsonString(xmlVedlegg.getTilleggsinfo()));
