+++ sendsoknad-business/src/test/java/no/nav/sbl/dialogarena/soknadinnsending/business/service/soknadservice/SoknadDataFletterTest.java
@@ -1,7 +1,6 @@
 package no.nav.sbl.dialogarena.soknadinnsending.business.service.soknadservice;
 
 import no.nav.melding.domene.brukerdialog.behandlingsinformasjon.v1.XMLHovedskjema;
-import no.nav.melding.domene.brukerdialog.behandlingsinformasjon.v1.XMLInnsendingsvalg;
 import no.nav.melding.domene.brukerdialog.behandlingsinformasjon.v1.XMLMetadataListe;
 import no.nav.melding.domene.brukerdialog.behandlingsinformasjon.v1.XMLVedlegg;
 import no.nav.sbl.dialogarena.sendsoknad.domain.DelstegStatus;
@@ -32,7 +31,6 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
-import org.mockito.Captor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
@@ -94,9 +92,9 @@ public class SoknadDataFletterTest {
     private MigrasjonHandterer migrasjonHandterer;
     @Mock
     private SkjemaOppslagService skjemaOppslagService;
+    @Mock
+    private LegacyInnsendingService legacyInnsendingService;
 
-    @Captor
-    private ArgumentCaptor<XMLHovedskjema> argument;
 
     @InjectMocks
     private SoknadDataFletter soknadServiceUtil;
@@ -121,6 +119,7 @@ public class SoknadDataFletterTest {
         soknadServiceUtil.alternativRepresentasjonService = alternativRepresentasjonService;
         when(config.hentStruktur(any(String.class))).thenReturn(new SoknadStruktur());
     }
+
 
     @Test
     public void skalStarteSoknad() {
@@ -180,47 +179,14 @@ public class SoknadDataFletterTest {
 
         when(lokalDb.hentSoknadMedVedlegg(behandlingsId)).thenReturn(webSoknad);
         when(lokalDb.hentSoknadMedData(1L)).thenReturn(webSoknad);
-        when(vedleggService.hentVedleggOgKvittering(webSoknad)).thenReturn(mockHentVedleggForventninger(webSoknad));
+//        when(vedleggService.hentVedleggOgKvittering(webSoknad)).thenReturn(mockHentVedleggForventninger(webSoknad));
         when(migrasjonHandterer.handterMigrasjon(any(WebSoknad.class))).thenReturn(webSoknad);
 
         soknadServiceUtil.sendSoknad(behandlingsId, new byte[]{1, 2, 3}, new byte[]{4,5,6});
 
-        verify(henvendelsesConnector).avsluttSoknad(eq(behandlingsId), argument.capture(),
-                refEq(new XMLVedlegg[] {
-                        new XMLVedlegg()
-                                .withUuid("uidVedlegg1")
-                                .withInnsendingsvalg(XMLInnsendingsvalg.LASTET_OPP.toString())
-                                .withFilnavn("Test Annet vedlegg")
-                                .withTilleggsinfo("Test Annet vedlegg")
-                                .withFilstorrelse("2")
-                                .withSideantall(3)
-                                .withMimetype("application/pdf")
-                                .withSkjemanummer("N6"),
-                                new XMLVedlegg()
-                                        .withInnsendingsvalg(XMLInnsendingsvalg.SENDES_IKKE.toString())
-                                        .withTilleggsinfo("")
-                                        .withSkjemanummer("L8")
-                                        .withFilnavn("L8"),
-                                new XMLVedlegg()
-                                        .withUuid("kvitteringRef")
-                                        .withInnsendingsvalg(XMLInnsendingsvalg.LASTET_OPP.toString())
-                                        .withFilnavn(SKJEMANUMMER_KVITTERING)
-                                        .withTilleggsinfo("")
-                                        .withFilstorrelse("3")
-                                        .withSideantall(1)
-                                        .withMimetype("application/pdf")
-                                        .withSkjemanummer(SKJEMANUMMER_KVITTERING)
-                })
-        , any());
-
-        XMLHovedskjema xmlHovedskjema = argument.getValue();
-        assertThat(xmlHovedskjema.getJournalforendeEnhet()).isEqualTo("enhet");
-        assertThat(xmlHovedskjema.getUuid()).isEqualTo("uidHovedskjema");
-        assertThat(xmlHovedskjema.getInnsendingsvalg()).isEqualTo(XMLInnsendingsvalg.LASTET_OPP.toString());
-        assertThat(xmlHovedskjema.getFilnavn()).isEqualTo(AAP+".pdfa");
-        assertThat(xmlHovedskjema.getFilstorrelse()).isEqualTo("3");
-        assertThat(xmlHovedskjema.getMimetype()).isEqualTo("application/pdf");
-        assertThat(xmlHovedskjema.getSkjemanummer()).isEqualTo(AAP);
+        verify(legacyInnsendingService, times(1)).sendSoknad(any(), any(), any());
+        verify(hendelseRepository, times(1)).hentVersjon(eq(behandlingsId));
+        verify(soknadMetricsService, times(1)).sendtSoknad(eq(AAP), eq(false));
     }
 
     @Test
